Optimize structure
